+++ src/main/java/com/example/quizmaster/service/QuizService.java
@@ -2,7 +2,7 @@ package com.example.quizmaster.service;
 
 import com.example.quizmaster.entity.*;
 import com.example.quizmaster.payload.ApiResponse;
-import com.example.quizmaster.payload.Pageable;
+import com.example.quizmaster.payload.CustomPageable;
 import com.example.quizmaster.payload.request.ReqPassTest;
 import com.example.quizmaster.payload.request.RequestQuiz;
 import com.example.quizmaster.payload.response.ResponseQuestion;
@@ -17,7 +17,6 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Collections;
@@ -34,6 +33,7 @@ public class QuizService {
     private final ResultRepository resultRepository;
     private final AnswerRepository answerRepository;
 
+    // categoriya saqlash
     public ApiResponse save(RequestQuiz requestQuiz) {
 
         Quiz quiz = Quiz.builder()
@@ -45,9 +45,10 @@ public class QuizService {
                 .build();
 
         quizRepository.save(quiz);
-        return new ApiResponse("Quiz saved successfully", HttpStatus.CREATED);
+        return new ApiResponse("Quiz saved successfully!", HttpStatus.CREATED);
     }
 
+    // categoriyalarni olish
     public ApiResponse getAll(int page, int size) {
         PageRequest pageRequest = PageRequest.of(page, size);
         Page<Quiz> quizPage = quizRepository.findAll(pageRequest);
@@ -56,7 +57,7 @@ public class QuizService {
                 .map(this::mapToResponseQuiz)
                 .collect(Collectors.toList());
 
-        Pageable pageable = Pageable.builder()
+        CustomPageable pageable = CustomPageable.builder()
                 .page(page)
                 .size(size)
                 .totalPage(quizPage.getTotalPages())
@@ -64,12 +65,15 @@ public class QuizService {
                 .data(responseQuizList)
                 .build();
 
-        return new ApiResponse("Quizzes retrieved successfully", HttpStatus.OK, pageable);
+        return new ApiResponse("Quizzes retrieved successfully!", HttpStatus.OK, pageable);
     }
 
+    // categoriyani yangilash
     public ApiResponse update(Long id, RequestQuiz requestQuiz) {
-        Quiz quiz = quizRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Quiz not found with id: " + id));
+        Quiz quiz = quizRepository.findById(id).orElse(null);
+        if (quiz == null) {
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
+        }
 
         quiz.setTitle(requestQuiz.getTitle());
         quiz.setDescription(requestQuiz.getDescription());
@@ -77,21 +81,25 @@ public class QuizService {
         quiz.setTimeLimit(requestQuiz.getTimeLimit());
 
         quizRepository.save(quiz);
-        return new ApiResponse("Quiz updated successfully", HttpStatus.OK);
+        return new ApiResponse("Quiz updated successfully!", HttpStatus.OK);
     }
 
+    // o'chirish
     public ApiResponse delete(Long id) {
-        Quiz quiz = quizRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Quiz not found with id: " + id));
+        Quiz quiz = quizRepository.findById(id).orElse(null);
+        if (quiz  == null){
+            return new ApiResponse("Quiz not found!",HttpStatus.NOT_FOUND);
+        }
 
         quizRepository.delete(quiz);
-        return new ApiResponse("Quiz deleted successfully", HttpStatus.OK);
+        return new ApiResponse("Quiz deleted successfully!", HttpStatus.OK);
     }
 
+    // categoriya boyicha random savollar chiqarish
     public ApiResponse getRandomQuestionsForQuiz(Long quiz) {
         Quiz quiz1 = quizRepository.findById(quiz).orElse(null);
         if (quiz1 == null) {
-            return new ApiResponse("Quiz not found", HttpStatus.NOT_FOUND);
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
         }
 
         List<Question> allQuestions = questionRepository.findAllByQuizId(quiz);
@@ -101,17 +109,20 @@ public class QuizService {
         int questionCount = quiz1.getQuestionCount();
 
         if (responseQuestions.size() > questionCount) {
-            return new ApiResponse("Success", HttpStatus.OK, responseQuestions.subList(0, questionCount));
+            return new ApiResponse("Success!", HttpStatus.OK, responseQuestions.subList(0, questionCount));
         }
 
-        return new ApiResponse("Insufficient number of questions available", HttpStatus.CONFLICT);
+        return new ApiResponse("Insufficient number of questions available!", HttpStatus.CONFLICT,responseQuestions);
     }
 
-    public ApiResponse passTest(List<ReqPassTest> passTestList, User user, Long quizId) {
+    // test yechish
+    public ApiResponse passTest(List<ReqPassTest> passTestList, User user, Long quizId,Long timeTaken) {
         Quiz quiz = quizRepository.findById(quizId).orElse(null);
         if (quiz == null) {
-            return new ApiResponse("Quiz not found", HttpStatus.NOT_FOUND);
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
         }
+
+        LocalDateTime startTime = LocalDateTime.now();
 
         List<Long> questionIds = passTestList.stream()
                 .map(ReqPassTest::getQuestionId)
@@ -124,23 +135,29 @@ public class QuizService {
                         .anyMatch(answer -> answer.getId().equals(reqPassTest.getAnswerId())))
                 .count();
 
+        LocalDateTime endTime = LocalDateTime.now();
+
+
         Result result = Result.builder()
                 .correctAnswers((int) correctCountAnswers)
                 .quiz(quiz)
                 .totalQuestion(passTestList.size())
                 .user(user)
+                .startTime(startTime)
+                .endTime(endTime)
+                .timeTaken(timeTaken)
                 .build();
 
         resultRepository.save(result);
-        return new ApiResponse("Test passed successfully", HttpStatus.OK);
+        return new ApiResponse("Test passed successfully!", HttpStatus.OK);
     }
 
-
-    public ApiResponse getOne(Long id){
+    // idi boyicha categoruiya qidirish
+    public ApiResponse getOne(Long id) {
         Quiz quiz = quizRepository.findById(id).orElse(null);
 
         if (quiz == null) {
-            return new ApiResponse("Quiz not found with this id: " + id, HttpStatus.NOT_FOUND);
+            return new ApiResponse("Quiz not found!", HttpStatus.NOT_FOUND);
         }
 
         ResponseQuiz responseQuiz = ResponseQuiz.builder()
@@ -152,10 +169,11 @@ public class QuizService {
                 .questionCount(quiz.getQuestionCount())
                 .build();
 
-        return new ApiResponse("Quiz found", HttpStatus.OK, responseQuiz);
+        return new ApiResponse("Quiz found!", HttpStatus.OK, responseQuiz);
 
     }
 
+    // quizni response qilish
     private ResponseQuiz mapToResponseQuiz(Quiz quiz) {
         return ResponseQuiz.builder()
                 .id(quiz.getId())
